Scraper0.1/requirements.txt
# Python package dependencies for Facebook scraper
# Core dependencies
selenium==4.10.0
webdriver-manager==3.9.1
sqlalchemy==2.0.17
psycopg2-binary==2.9.6
python-dotenv==1.0.0

# Data processing and utilities
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
tqdm==4.65.0
python-dateutil==2.9.0.post0
dateparser==1.2.1

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3

# Image processing
pillow==10.0.0

# Command line interface
tabulate==0.9.0
colorama==0.4.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
